hrm_eval/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.3.0
pytest-cov>=4.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
httpx>=0.25.0
pytest-mock>=3.11.0
tensorboard>=2.13.0
wandb>=0.15.0
loguru>=0.7.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
openai>=1.0.0
anthropic>=0.7.0
pinecone-client>=2.2.0

